refactor: moves types to dedicated file
